fix layer scale
